Library can now publish / subscribe to messages greater than 127 bytes in length. DEV-1 #resolve

+++ c_embedded/client/example.c
@@ -166,7 +166,7 @@ int read_packet(int timeout) {
 
 
 int main(int argc, char* argv[]) {
-	int packet_length;
+	int16_t packet_length;
 	//uint16_t msg_id, msg_id_rcv;
 	mqtt_broker_handle_t broker;
